requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
Faker==24.3.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.1
